Makefile: Always build static binaries
Static binaries were use for arm64 through the cross-compilation but not
for amd64. This added a dependency on some system libraries that the
host OS should provide, such as glibc with a certain minimal version.

To align the amd64 with the arm64 binary and reduce requirements for the
host OS, always build static binaries. The result was checked with
running "file bin/applicationhealth-*".