Resource governance, heartbeat and dev container changes
The main feature change here is the addition of resource governance for linux via cgroups.

We discover the current cgroup and add a sub cgroup for our purposes (limiting cpu to 1% and memory to 40MB)

I also added support for detecting a stuck vmwatch using the heartbeat file and implemented the same logic for restarts from the windows version (3 restarts per 3 hours)

As part of the development of this, I added support for devcontainer execution so we can step through the code from a dev machine into either a WSL session or a linux vm with tools installed.

I added integration tests to check process exit, OOM and cpu throttling.  These changes required a few changes to the makefile and scripts.

I also updated the vmwatch binaries and added a script to download the latest ones as well

I updated the govendor files using the tool it told me to run I hope I did this right