Fixing CodeQL error for windows Builds (#82)
This pull request primarily focuses on extending the CodeQL analysis to
the Windows environment in the `.github/workflows/codeql.yml` file. The
changes include adding `windows-latest` to the operating systems list
and adjusting the build commands to accommodate this new environment.

Key changes:

*
[`.github/workflows/codeql.yml`](diffhunk://#diff-12783128521e452af0cfac94b99b8d250413c516ec71fe6d97dbea666ff7ba27L57-R57):
Added `windows-latest` to the list of operating systems under the
`jobs:` section, allowing CodeQL analysis to run on both Ubuntu and
Windows environments.
*
[`.github/workflows/codeql.yml`](diffhunk://#diff-12783128521e452af0cfac94b99b8d250413c516ec71fe6d97dbea666ff7ba27L95-R102):
Modified the conditions for running build commands. If the build mode is
'manual' and the operating system is 'ubuntu-latest', the usual `go env`
and `make binary` commands are run. If the operating system is
'windows-latest', `go env` and `make binary-windows` commands are run
instead.