chore(logging): Logging the error code if kill fails
We investigated this at length and cannot figure out what could possibly be happening based on the logs.  Logging the fact that kill failed should help us at least narrow it down further.

Also adding a sleep between killing app health and vmwatch to reduce the chance of any races (vmwatch should die naturally when app health is killed but it may not happen immediately so giving it a second to respond should reduce the times it needs to be killed independently.